Linear perturbation around steady state

Note

The main functions of this section are in the folder LinearizationFunctions.

The model is linearized with respect to aggregate variables. For this, we write the equilibrium conditions in the form of $F(X,X')=0$, where $X$ and $X'$ are (expected) deviations from steady state in two successive periods. Applying the total differential yields $A*X' = - B*X$, where $A$,$B$ are the first derivatives of $F$ with respect to $X'$,$X$. In the standard setting, we use the generalized Schur decomposition [Klein] to transform this equation into a linearized observation equation $d = gx*k$ and a linearized state transition equation $k' = hx*k$, where $k$ is a vector of the state variables and $d$ is a vector of the control variables ($X = \begin{bmatrix} k \\ d \end{bmatrix}$).

In our code, $F$ is implemented as Fsys(), while differentiating and solving for $gx$ and $hx$ is done in SGU(), and linearize_full_model() returns the results as a struct LinearResults:

HANKEstim.linearize_full_model — Function

linearize_full_model()

Linearize the full solution (i.e. with idiosyncratic states and controls) around the steady state, using SGU().

Returns

struct LinearResults, containing

A::Array{Float64,2},B::Array{Float64,2}: first derivatives of Fsys() with respect to arguments X [B] and XPrime [A]

State2Control::Array{Float64,2}: observation equation

LOMstate::Array{Float64,2}: state transition equation

Overview of SGU()

HANKEstim.SGU — Function

SGU(XSS,A,B,m_par,n_par,indexes,Copula,compressionIndexes,distrSS;estim)

Calculate the linearized solution to the non-linear difference equations defined by function Fsys(), using Schmitt-Grohé & Uribe (JEDC 2004) style linearization (apply the implicit function theorem to obtain linear observation and state transition equations).

The Jacobian is calculated using dual numbers (implemented by package ForwardDiff). Use macro @make_deriv to compute partials simultaneously, with chunk size given by global n_FD. Make use of model knowledge to set some entries manually.

Arguments

XSS: steady state around which the system is linearized

A,B: matrices to be filled with first derivatives (see Returns)

m_par::ModelParameters, n_par::NumericalParameters: n_par.sol_algo determines the solution algorithm

Copula::Function,distrSS::Array{Float64,3}: Copula maps marginals to linearized approximation of joint distribution around distrSS

indexes::IndexStruct,compressionIndexes: access states and controls by name (DCT coefficients of compressed $V_m$ and $V_k$ in case of compressionIndexes)

Returns

gx,hx: observation equations [gx] and state transition equations [hx]

alarm_sgu,nk: alarm_sgu=true when solving algorithm fails, nk number of predetermined variables

A,B: first derivatives of Fsys() with respect to arguments X [B] and XPrime [A]

The function executes the following steps:

generate devices to retrieve distribution and marginal value functions from compressed states/controls (Γ and DC,IDC)

calculate the first derivative of Fsys() with respect to X and XPrime. We use automatic differentiation (implemented in Julia by the package ForwardDiff). Partial derivatives are calculated in parallel, with chunk size n_FD (see @make_deriv). Additionally, we manually set the partial derivatives of contemporaneous marginal value functions and the future marginal distributions [BL].

Note

n_FD is set as a global variable in the main script HANKEstim.jl.

compute linear observation and state transition equations. The solution algorithm is set in n_par.sol_algo, with the options :schur (mentioned above) and :lit [lit]. The results are matrices that map contemporaneous states to controls [gx], or contemporaneous states to future states [hx]

Called functions / macros

HANKEstim.@make_deriv — Macro

@make_deriv(n_FD_s)

Set dual number seeds at chunk of size n_FD_s (starting at index i) of the argument X (XPrime) in Fsys() and save matrix of partial derivatives (for this group of variables) in Deriv(i) (DerivPrime(i)).

Require length_X0 and remaining arguments of Fsys() in namespace.

Overview of Fsys()

HANKEstim.Fsys — Function

Fsys(X,XPrime,Xss,m_par,n_par,indexes,Γ,compressionIndexes,DC,IDC)

Equilibrium error function: returns deviations from equilibrium around steady state.

Split computation into Aggregate Part, handled by Fsys_agg(), and Heterogeneous Agent Part.

Arguments

X,XPrime: deviations from steady state in periods t [X] and t+1 [XPrime]

Xss: states and controls in steady state

Γ,DC,IDC: transformation matrices to retrieve marginal distributions [Γ] and marginal value functions [DC,IDC] from deviations

indexes,compressionIndexes: access Xss by variable names (DCT coefficients of compressed $V_m$ and $V_k$ in case of compressionIndexes)

Example

julia> # Solve for steady state, construct Γ,DC,IDC as in SGU()
julia> Fsys(zeros(ntotal),zeros(ntotal),XSS,m_par,n_par,indexes,Γ,compressionIndexes,DC,IDC)
*ntotal*-element Array{Float64,1}:
 0.0
 0.0
 ...
 0.0

The function Fsys() proceeds in the following way:

set up vector F, that contains the errors to all equilibrium conditions. There are as many conditions as deviations from steady state (length of X,XPrime), and conditions are indexed with respective model variable in IndexStruct indexes

generate locally all aggregate variables (for both periods) using @generate_equations

construct the distributions and the marginal value functions from the steady state values and the (compressed) deviations

write all equilibrium condition-errors with respect to aggregate variables to F, using Fsys_agg()

compute optimal policies with EGM_policyupdate(), given future marginal value functions, prices, and individual incomes. Infer present marginal value functions from them (envelope theorem) and set the difference to assumed present marginal value functions (in terms of their compressed deviation from steady state) as equilibrium condition-errors (backward iteration of the value function)

compute future distribution from previous distribution and optimal asset policies. Set difference to assumed future marginal distributions as equilibrium condition-errors (forward iteration of the distribution)

compute distribution summary statistics with distrSummaries() and write equilibrium conditions with their respective (control) variables

return F

Called functions / macros

HANKEstim.@generate_equations — Macro

@generate_equations(var_names)

Write out the expansions around steady state for all variables in var_names, i.e. generate code that reads aggregate states/controls from steady state deviations.

Equations take the form of (with variable r as example):

r = exp.(Xss[indexes.rSS] .+ X[indexes.r])

rPrime = exp.(Xss[indexes.rSS] .+ XPrime[indexes.r])

HANKEstim.Fsys_agg — Function

Fsys_agg(X,XPrime,Xss,distrSS,m_par,n_par,indexes)

Return deviations from aggregate equilibrium conditions.

indexes can be both IndexStruct or IndexStructAggr; in the latter case (which is how function is called by SGU_estim()), variable-vectors X,XPrime, and Xss only contain the aggregate variables of the model.

KleinSee the paper Using the generalized Schur form to solve a multivariate linear rational expectations model by Paul Klein (JEDC 2000)

BLContemporaneous marginal value functions are irrelevant for optimal decisions, so its effect on other model variables is 0. Due to a rich enough set of prices, the future distribution directly only affects the Fokker-Planck equation. For details, see the paper Solving heterogeneous agent models in discrete time with many idiosyncratic states by perturbation methods

litInvoking the Implicit Function Theorem, there exist functions $g$ and $h$ such that $F\left(\begin{pmatrix} k \\ g(k) \end{pmatrix},\begin{pmatrix} h(k) \\ g(h(k)) \end{pmatrix}\right)=0$. Totally differentiating by $k$ yields $B \begin{pmatrix}\mathbb{I}\\ Dg \end{pmatrix}+A \begin{pmatrix}\mathbb{I}\\ Dg \end{pmatrix} Dh = 0$. The :lit-algorithm solves this equation for $Dg$ and $Dh$ iteratively.